LayoutTests/js/dom/array-sort-non-configurable-element-expected.txt

PASS array.sort() threw exception TypeError: Unable to delete property..
PASS successfullyParsed is true

TEST COMPLETE
